Removed redundant code and optimized Match model
