+++ cricket/views/__init__.py
@@ -1,0 +1,11 @@
+# Import all views for easy access
+from .auth_views import signup_view, login_view, logout_view
+from .match_views import (
+    match_list, add_match, match_details, delete_match, delete_all_matches,
+    start_game, live_match, get_live_match_data, update_match
+)
+from .player_views import player_list, add_player
+from .game_flow_views import (
+    set_bowler, select_next_bowler, select_next_batter, handle_wicket
+)
+from .base_views import home, clear_message_flag
