+++ cricket/views/auth_views.py
@@ -1,0 +1,37 @@
+from django.shortcuts import render, redirect
+from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
+from django.contrib.auth import authenticate, login, logout
+from django.contrib import messages
+
+def signup_view(request):
+    """Handle user registration."""
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            form.save()
+            messages.success(request, "Signup successful! Please log in.")
+            return redirect('login')  # Redirect to login after signup
+    else:
+        form = UserCreationForm()
+    return render(request, 'cricket/signup.html', {'form': form})
+
+def login_view(request):
+    """Handle user login."""
+    if request.method == 'POST':
+        form = AuthenticationForm(data=request.POST)
+        if form.is_valid():
+            user = form.get_user()
+            login(request, user)
+            # Redirect to 'next' parameter if available, else home
+            next_url = request.GET.get('next', 'home')
+            return redirect(next_url)
+        else:
+            messages.error(request, "Invalid credentials.")
+    else:
+        form = AuthenticationForm()
+    return render(request, 'cricket/login.html', {'form': form})
+
+def logout_view(request):
+    """Handle user logout."""
+    logout(request)
+    return redirect('home')  # Redirect to homepage after logout
